pyproject.toml
[tool.poetry]
name = "trading-ig"
version = "0.0.23"
description = "A lightweight Python wrapper for the IG Markets API"
authors = ["Femto Trader <femto.trader@gmail.com>", "Andy Geach <andy@bugorfeature.net>"]
maintainers = ["Andy Geach <andy@bugorfeature.net>"]
license = "BSD-3-Clause"
readme = "README.rst"
homepage = "https://github.com/ig-python/trading-ig"
repository = "https://github.com/ig-python/trading-ig"
documentation = "https://trading-ig.readthedocs.io/"
keywords = ["trading", "spread betting", "CFDs"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Cython",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: BSD License",
]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.24"
pycryptodome = "^3.9"
requests-cache = "^0.5"
six = "^1.15"
lightstreamer-client-lib = "^1.0.3"

pandas = {version = "^2", optional = true}
munch = {version = "^2.5", optional = true}
tenacity = {version = "^8", optional = true}
redis = {version = "^6", optional = true}
psycopg2-binary = {version = "^2.9", optional = true}

[tool.poetry.extras]
pandas = ["pandas"]
munch = ["munch"]
tenacity = ["tenacity"]
sphinx = ["sphinx"]
docs = [
    "sphinx"
]
trading-board = ["redis", "psycopg2-binary"]

[tool.poetry.group.dev.dependencies]
flake8 = "^7"
pytest = "^8"
responses = "^0.25"
coveralls = "^3"
sphinx-rtd-theme = {version = "^1.0.0", optional = true}
black = "23.3.0"

[tool.poetry.urls]
"Issues" = "https://github.com/ig-python/trading-ig/issues"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
